Added a Python-style config processor and expander.
Using a library of predefined GCE Resource constructors, we can simplify the
computation and expansion of configurations by allowing the user to use the
power of the Python interpreter to construct their objects.

This allows the full spectrum of computation from Python, including function
calls, string concatenation, list comprehension, etc.

For end-user simplicity, we compile + evaluate the user's config module in an
environment where we have already imported every top-level definition from the
`gce` module, thus allowing the user to immediately start writing the config
without having to remember to import a module that they don't need to know
about, since it's boilerplate to be added to the top of every config.