Removed the implicit default config of `ambari.yaml`.
This is problematic because a user might not be expecting to bring up an Ambari
cluster as part of running the tool, and also, Ambari config requires running
`make` to generate the startup scripts first, so this avoids confusing errors.

From now on, users will have to explicitly specify which config they want to
run, which will make even more sense as we build up a library of configs and
having it default to Ambari will become even more perplexing.