Switch from `execfile()` to `imp.load_source()`.
This enables us to have files which have `import` statements, but because it
disallows us to inject custom imports before loading, it comes at the cost of
forcing users to add `import gce` or `from gce import *` explicitly to the top
of their files. As a side-effect, this enables a multi-cloud strategy.