Link to the directory rather than the README.
GitHub automatically displays the README in a directory, but linking to the
directory lets the user also see the contained files and directories within.

Linking to a specific file (especially README) is confusing as there is no
navigation once the user is looking at the page, and one needs to explicitly
navigate to the directory listing to look at the files/directories further.