Clarify browser config requirement for proxy.
Previously, the need to configure a browser for use with the proxy was buried in
the bullet for starting up the proxy and thus is easiy to gloss over when you
can just copy-paste the URL and get started.

Thus, users would get error messages about inability to resolve the URL because
they did not update their browser's proxy settings with no indication of what to
do to fix the error.