Switch to single Ambari image for agent + server.
Previously, we were building separate GCE VM images for both Ambari agent and
server. With this change, we build a single image which includes both agent and
server, both disabled and neither one automatically starting on boot.

Then, we choose which mode the VM will start in via the startup script.

This lets us speed up and simplify builds as well as decrease storage costs.