Break up ssh command lines to support Windows.
The ssh command that comes with PuTTY on Windows cannot handle flags with spaces
in them, nor can it handle when those flags and their options are concatenated,
so to easily support all operating systems with a single command, we are
breaking them up.

Also, to generalize beyond *nix-style OSes, removed the `${...}` notation to
avoid suggeting environment variables by that name and instead use placeholders
formatted as `{{...}}` in their place. Updated commands and instructions to use
the new format.