Added support for Travis CI.
Using `language: generic` rather than explicitly Python as it avoids a problem
with the OS X build which is having trouble installing Python 2.7.x (which
appears to already exist on the system), even with `sudo: yes`.

Not using `sudo` also allows us to run on the faster container-based
infrastructure, which reduces the latency of each test cycle.