Add Python 3 testing on macOS
Since we always use `#!/usr/bin/python` in our Python scripts, and since
`/usr/bin/python` on all version of macOS is Python 2.7.x, we are not
actually testing with Python3 regardless of the version of our Xcode
installation. This attempt sets an explicit choice to use Python3 for
several of macOS versions, and the test runner script is adjusted to
take into account the env var `$PYTHON`, which, if set, points to the
preferred Python binary to use for running the Python tests.

Also ensured code is now Python 3 compatible:

* replaced `print` statement with `print()` function
* replaced `dict.iteritems()` with `dict.items()`